+++ RenderOpenGL/Renderer.h
@@ -84,9 +84,9 @@ namespace Engine {
 
         //glm::vec3 cameraUp = glm::vec3(0.0f, 1.0f, 0.0f);
 
-        float deltaTime = 0.0f;
+        double deltaTime = 0.0f;
 
-        float lastFrame = 0.0f;
+        double lastFrame = 0.0f;
 
 		friend class InputManager;
